move useful coroutine utils to common utils module, update JavaTool.execute to respect coroutine context to avoid deadlocks
