+++ agents/agents-core/src/jvmCommonMain/kotlin/ai/koog/agents/core/agent/entity/AIAgentStrategy.kt
@@ -3,17 +3,19 @@
 package ai.koog.agents.core.agent.entity
 
 import ai.koog.agents.core.agent.context.AIAgentContext
-import ai.koog.agents.core.utils.asCoroutineContext
-import kotlinx.coroutines.runBlocking
+import ai.koog.utils.annotations.InternalKoogUtils
+import ai.koog.utils.coroutines.asCoroutineContext
+import ai.koog.utils.coroutines.runBlockingIfRequired
 import java.util.concurrent.ExecutorService
 
+@OptIn(InternalKoogUtils::class)
 public actual interface AIAgentStrategy<TInput, TOutput, TContext : AIAgentContext> {
     public actual val name: String
 
     public actual suspend fun execute(context: TContext, input: TInput): TOutput?
 
     public fun execute(context: TContext, input: TInput, executorService: ExecutorService? = null): TOutput? =
-        runBlocking(executorService.asCoroutineContext()) {
+        runBlockingIfRequired(executorService.asCoroutineContext()) {
             execute(context, input)
         }
 }
